worked on the changes to the login to look better and match the signup section along with the about us
