--- client/src/components/about.jsx
@@ -1,230 +1,0 @@
-import React from "react";
-const people = [
-  {
-    name: "Ryan Hemlock",
-    role: "Co-Founder / CEO",
-    imageUrl: "",
-  },
-  {
-    name: "Mustapha Kromah",
-    role: "Co-Founder / CEO",
-    imageUrl:
-      "https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
-  },
-  {
-    name: "Mykhalio Zakharkiv",
-    role: "Co-Founder / CEO",
-    imageUrl:
-      "https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
-  },
-];
-
-export default function About() {
-  return (
-    <div className="bg-white py-24 sm:py-32">
-      <div className="mx-auto grid max-w-7xl gap-x-8 gap-y-20 px-6 lg:px-8 xl:grid-cols-3">
-        <div className="max-w-2xl">
-          <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
-            Meet our leadership
-          </h2>
-          <p className="mt-6 text-lg leading-8 text-gray-600">
-            Libero fames augue nisl porttitor nisi, quis. Id ac elit odio vitae
-            elementum enim vitae ullamcorper suspendisse.
-          </p>
-        </div>
-        <ul
-          role="list"
-          className="grid gap-x-8 gap-y-12 sm:grid-cols-2 sm:gap-y-16 xl:col-span-2"
-        >
-          {people.map((person) => (
-            <li key={person.name}>
-              <div className="flex items-center gap-x-6">
-                <img
-                  className="h-16 w-16 rounded-full"
-                  src={person.imageUrl}
-                  alt=""
-                />
-                <div>
-                  <h3 className="text-base font-semibold leading-7 tracking-tight text-gray-900">
-                    {person.name}
-                  </h3>
-                  <p className="text-sm font-semibold leading-6 text-indigo-600">
-                    {person.role}
-                  </p>
-                </div>
-              </div>
-            </li>
-          ))}
-        </ul>
-      </div>
-      <div className="flex items-stretch justify-center">
-        <div className="grid md:grid-cols-2">
-          <div className="h-full pr-6">
-            <p className="mt-3 mb-12 text-lg text-gray-800">
-              Lorem, ipsum dolor sit amet consectetur adipisicing elit. Est
-              aperiam dolore facilis nulla dignissimos quasi accusamus,
-              temporibus recusandae, ab inventore quibusdam dolor beatae
-              corporis ea consectetur id ad nam velit!
-            </p>
-            <ul className="mb-6 md:mb-0">
-              <li className="flex">
-                <div className="flex h-10 w-10 items-center justify-center rounded bg-green-600 text-gray-50">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="24"
-                    height="24"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    className="h-6 w-6"
-                  >
-                    <path d="M9 11a3 3 0 1 0 6 0a3 3 0 0 0 -6 0"></path>
-                    <path d="M17.657 16.657l-4.243 4.243a2 2 0 0 1 -2.827 0l-4.244 -4.243a8 8 0 1 1 11.314 0z"></path>
-                  </svg>
-                </div>
-                <div className="ml-4 mb-4">
-                  <h3 className="mb-2 text-lg font-medium leading-6 text-gray-800">
-                    Our Address
-                  </h3>
-                  <p className="text-gray-600 dark:text-slate-600">
-                    Philadelphia
-                  </p>
-                  <p className="text-gray-600 dark:text-slate-600">
-                    Pennsylvania, United States
-                  </p>
-                </div>
-              </li>
-              <li className="flex">
-                <div className="flex h-10 w-10 items-center justify-center rounded bg-green-600 text-gray-50">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="24"
-                    height="24"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    className="h-6 w-6"
-                  >
-                    <path d="M5 4h4l2 5l-2.5 1.5a11 11 0 0 0 5 5l1.5 -2.5l5 2v4a2 2 0 0 1 -2 2a16 16 0 0 1 -15 -15a2 2 0 0 1 2 -2"></path>
-                    <path d="M15 7a2 2 0 0 1 2 2"></path>
-                    <path d="M15 3a6 6 0 0 1 6 6"></path>
-                  </svg>
-                </div>
-                <div className="ml-4 mb-4">
-                  <h3 className="mb-2 text-lg font-medium leading-6 text-gray-800">
-                    Contact
-                  </h3>
-                  <p className="text-gray-600 dark:text-slate-600">
-                    Mobile: +1 (215) 898-5000
-                  </p>
-                  <p className="text-gray-600 dark:text-slate-600">
-                    Mail: customerservice@
-                    <a href="/" className="text-green-600">
-                      Project 3
-                    </a>
-                    .com
-                  </p>
-                </div>
-              </li>
-              <li className="flex">
-                <div className="flex h-10 w-10 items-center justify-center rounded bg-green-600 text-gray-50">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="24"
-                    height="24"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    className="h-6 w-6"
-                  >
-                    <path d="M3 12a9 9 0 1 0 18 0a9 9 0 0 0 -18 0"></path>
-                    <path d="M12 7v5l3 3"></path>
-                  </svg>
-                </div>
-                <div className="ml-4 mb-4">
-                  <h3 className="mb-2 text-lg font-medium leading-6 text-gray-800">
-                    Working hours
-                  </h3>
-                  <p className="text-gray-600 dark:text-slate-600">
-                    Monday - Friday: 08:00 - 17:00
-                  </p>
-                  <p className="text-gray-600 dark:text-slate-600">
-                    Saturday &amp; Sunday: 08:00 - 12:00
-                  </p>
-                </div>
-              </li>
-            </ul>
-          </div>
-          <div className="card h-fit max-w-6xl p-5 md:p-12" id="form">
-            <h2 className="mb-4 text-2xl font-bold">Ready to Get Started?</h2>
-            <form id="contactForm">
-              <div className="mb-6">
-                <div className="mx-0 mb-1 sm:mb-4">
-                  <div className="mx-0 mb-1 sm:mb-4">
-                    <label
-                      htmlFor="name"
-                      className="pb-1 text-xs uppercase tracking-wider"
-                    ></label>
-                    <input
-                      type="text"
-                      id="name"
-                      autoComplete="given-name"
-                      placeholder="Your name"
-                      className="mb-2 w-full rounded-md border focus:ring-green-600 focus:border-green-600 border-gray-400 py-2 pl-2 pr-4 shadow-md dark:text-gray-300 sm:mb-0"
-                      name="name"
-                    ></input>
-                  </div>
-                  <div className="mx-0 mb-1 sm:mb-4">
-                    <label
-                      htmlFor="email"
-                      className="pb-1 text-xs uppercase tracking-wider"
-                    ></label>
-                    <input
-                      type="email"
-                      id="email"
-                      autoComplete="email"
-                      placeholder="Your email address"
-                      className="mb-2 w-full rounded-md border focus:ring-green-600 focus:border-green-600 border-gray-400 py-2 pl-2 pr-4 shadow-md dark:text-gray-300 sm:mb-0"
-                      name="email"
-                    ></input>
-                  </div>
-                </div>
-                <div className="mx-0 mb-1 sm:mb-4">
-                  <label
-                    htmlFor="textarea"
-                    className="pb-1 text-xs uppercase tracking-wider"
-                  ></label>
-                  <textarea
-                    id="textarea"
-                    name="textarea"
-                    cols="30"
-                    rows="5"
-                    placeholder="Write your message..."
-                    className="mb-2 w-full rounded-md border focus:ring-green-600 focus:border-green-600 border-gray-400 py-2 pl-2 pr-4 shadow-md sm:mb-0"
-                  ></textarea>
-                </div>
-              </div>
-              <div className="text-center">
-                <button
-                  type="submit"
-                  className="w-full bg-green-600 text-white px-6 py-3 font-xl rounded-md hover:bg-green-500 sm:mb-0"
-                >
-                  Send Message
-                </button>
-              </div>
-            </form>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
